// src/types/api.ts
export interface AiArticle {
  _id: string;
  title: string;
  description: string;
  category: string[];
  seo: {
    meta_title: string;
    meta_description: string;
    keywords: string[];
  };
  slug: string;
  content: string;
  summary: string;
  leagueName: string | null;
  leagueId: string | null;
  imageUrls: string[];
  scheduledTime: string | null;
  targetWebsites: string[];
  createdAt: string;
  updatedAt: string;
}

export interface AiArticleResponse {
  data: AiArticle[];
  meta: {
    currentPage: number;
    totalPages: number;
    totalFeeds: number;
    limit: number;
  };
}